Update project for Xcode 8 and share scheme

+++ MultiSegment/ViewController.m
@@ -48,11 +48,12 @@
 - (IBAction)selectNone {
     [self.multiSelectControl selectAllSegments:NO];
 }
--(void)multiSelect:(MultiSelectSegmentedControl *)multiSelecSegmendedControl didChangeValue:(BOOL)value atIndex:(NSUInteger)index{
+
+- (void)multiSelect:(MultiSelectSegmentedControl *)multiSelecSegmendedControl didChangeValue:(BOOL)value atIndex:(NSUInteger)index{
     if (value) {
-        NSLog(@"multiSelect with tag %i selected button at index: %i", multiSelecSegmendedControl.tag, index);
+        NSLog(@"multiSelect with tag %li selected button at index: %lu", (long)multiSelecSegmendedControl.tag, (unsigned long)index);
     } else {
-        NSLog(@"multiSelect with tag %i deselected button at index: %i", multiSelecSegmendedControl.tag, index);
+        NSLog(@"multiSelect with tag %li deselected button at index: %lu", (long)multiSelecSegmendedControl.tag, (unsigned long)index);
     }
 }
